photo with comments
These comments are available: 

The communications physical layer consists of a single BMS connected to multiple BMs. 

The communication is via a multi-drop +5V level TTL serial UART bus. The output from the BMS is connected in common to the input pins of all attached BMs. All of the outputs from the BMs are wired-or connected together to the input of the BMS. All BM outputs have series diodes and can only pull down. 

If you put 5v between VCC and ground, it lights up optocoupler IC3 which tells the buck converter on the BMS to generate 3.3V to power the CPU (drawing from the 40V from the 12S battery bank)

(Note: To communicate without a BMS, a pull-up resistor is required on the outputs from the BMs.) At the writing this document, latest release (V1_1C) use baud rate of 115200. In the next release, the serial baud rate will be 230400. 8 data bits, 1 stop bit, no parity are used.